requirements.txt
# --- Core do Bot ---
discord.py
python-dotenv

# --- Banco de Dados (Versão estável sem PyIceberg) ---
supabase<2.10.0
gotrue<3.0.0
postgrest<0.18.0
storage3<0.9.0
supafunc<0.7.0

# --- Dependências de Conexão e Segurança ---
websockets<16
pyjwt
cryptography
strenum
anyio
httpcore
h11
h2
sniffio

# --- Tipagem e Validação ---
pydantic
pydantic-core
annotated-types
typing-extensions
typing-inspection
deprecation

# --- Remoção de Fundo (Leve para PC fraco) ---
rembg[cpu]
onnxruntime



#.\venv\Scripts\activate